requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
colorlog==6.6.0
frozenlist==1.3.0
geographiclib==1.52
geopy==2.2.0
hikari==2.0.0.dev109
hikari-lightbulb==2.2.2
idna==3.3
multidict==6.0.2
requests==2.28.1
urllib3==1.26.9
yarl==1.7.2
